+++ samples/My.Hr/My.Hr.Test/Apis/EmployeeTest.cs
@@ -11,17 +11,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 
-namespace My.Hr.Test
+namespace My.Hr.Test.Apis
 {
     [TestFixture, NonParallelizable]
-    public class EmployeeTest : UsingAgentTesterServer<Startup>
+    public class EmployeeTest
     {
         #region Get
 
         [Test, TestSetUp]
         public void A110_Get_NotFound()
         {
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.GetAsync(404.ToGuid()));
         }
@@ -29,7 +31,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A120_Get_Found_NoAddress()
         {
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .IgnoreChangeLog()
                 .IgnoreETag()
@@ -50,7 +54,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A120_Get_Found_WithAddress()
         {
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .IgnoreChangeLog()
                 .IgnoreETag()
@@ -73,13 +79,15 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A120_Get_Modified_NotModified()
         {
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(1.ToGuid(), new WebApiRequestOptions { ETag = TestSetUp.ConcurrencyErrorETag })).Value!;
 
             Assert.NotNull(v);
 
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.NotModified)
                 .Run(a => a.GetAsync(1.ToGuid(), new WebApiRequestOptions { ETag = v.ETag }));
         }
@@ -91,7 +99,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A210_GetByArgs_All()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(null)).Value;
 
@@ -104,7 +114,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A210_GetByArgs_All_Paging()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { IsIncludeTerminated = true }, PagingArgs.CreateSkipAndTake(1,2))).Value;
 
@@ -117,7 +129,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A220_GetByArgs_FirstName()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { FirstName = "*a*" })).Value;
 
@@ -130,7 +144,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A230_GetByArgs_LastName()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { LastName = "s*" })).Value;
 
@@ -143,7 +159,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A230_GetByArgs_LastName_IncludeTerminated()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { LastName = "s*", IsIncludeTerminated = true })).Value;
 
@@ -156,7 +174,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A240_GetByArgs_Gender()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { Genders = new List<string> { "F" } })).Value;
 
@@ -169,7 +189,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A250_GetByArgs_Empty()
         {
-            var v = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { LastName = "s*", FirstName = "b*", Genders = new List<string> { "F" } })).Value;
 
@@ -181,7 +203,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A260_GetByArgs_FieldSelection()
         {
-            var r = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { Genders = new List<string> { "F" } }, requestOptions: new WebApiRequestOptions().Include("firstname", "lastname")));
 
@@ -196,7 +220,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A270_GetByArgs_RefDataText()
         {
-            var r = AgentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<EmployeeAgent, EmployeeBaseCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByArgsAsync(new EmployeeArgs { Genders = new List<string> { "F" } }, requestOptions: new WebApiRequestOptions { IncludeRefDataText = true }));
 
@@ -215,6 +241,8 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void B110_Create()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             var v = new Employee
             {
                 Email = "j.smith@org.com",
@@ -229,7 +257,7 @@ namespace My.Hr.Test
             };
 
             // Create value.
-            v = AgentTester.Test<EmployeeAgent, Employee>()
+            v = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.Created)
                 .ExpectChangeLogCreated()
                 .ExpectETag()
@@ -239,7 +267,7 @@ namespace My.Hr.Test
                 .Run(a => a.CreateAsync(v)).Value!;
 
             // Check the value was created properly.
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(v.Id));
@@ -248,6 +276,8 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void B120_Create_Duplicate()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             var v = new Employee
             {
                 Email = "w.jones@org.com",
@@ -260,7 +290,7 @@ namespace My.Hr.Test
             };
 
             // Create value.
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.Conflict)
                 .Run(a => a.CreateAsync(v));
         }
@@ -272,13 +302,15 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C110_Update_NotFound()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(4.ToGuid())).Value!;
 
             // Try updating with an invalid identifier.
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.UpdateAsync(v, 404.ToGuid()));
         }
@@ -286,20 +318,22 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C120_Update_Concurrency()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 4.ToGuid();
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
             // Try updating the value with an invalid eTag (if-match).
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.UpdateAsync(v, id, new WebApiRequestOptions { ETag = TestSetUp.ConcurrencyErrorETag }));
 
             // Try updating the value with an invalid eTag.
             v.ETag = TestSetUp.ConcurrencyErrorETag;
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.UpdateAsync(v, id));
         }
@@ -307,9 +341,11 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C130_Update()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 4.ToGuid();
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
@@ -321,7 +357,7 @@ namespace My.Hr.Test
             v.EmergencyContacts.Add(new EmergencyContact { FirstName = "Danny", LastName = "Keen", PhoneNo = "(234) 297 9834", Relationship = "FRD" });
 
             // Update the value.
-            v = AgentTester.Test<EmployeeAgent, Employee>()
+            v = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectChangeLogUpdated()
                 .ExpectETag(v.ETag)
@@ -331,7 +367,7 @@ namespace My.Hr.Test
                 .Run(a => a.UpdateAsync(v, id)).Value!;
 
             // Check the value was updated properly.
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(id));
@@ -340,9 +376,11 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C140_Update_AlreadyTerminated()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 2.ToGuid();
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
@@ -351,7 +389,7 @@ namespace My.Hr.Test
             v.LastName += "Y";
 
             // Update the value.
-            var r = AgentTester.Test<EmployeeAgent, Employee>()
+            var r = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.BadRequest)
                 .Run(a => a.UpdateAsync(v, id));
 
@@ -365,13 +403,15 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void D110_Patch_NotFound()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(4.ToGuid())).Value!;
 
             // Try patching with an invalid identifier.
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.PatchAsync(WebApiPatchOption.MergePatch, "{ \"lastName\": \"Smithers\" }", 404.ToGuid()));
         }
@@ -379,20 +419,22 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void D120_Patch_Concurrency()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 4.ToGuid();
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
             // Try updating the value with an invalid eTag (if-match).
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.PatchAsync(WebApiPatchOption.MergePatch, "{ \"lastName\": \"Smithers\" }", id, new WebApiRequestOptions { ETag = TestSetUp.ConcurrencyErrorETag }));
 
             // Try updating the value with an eTag header (json payload eTag is ignored).
             v.ETag = TestSetUp.ConcurrencyErrorETag;
-            AgentTester.Test<EmployeeAgent, Employee>()
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.PatchAsync(WebApiPatchOption.MergePatch, "{{ \"lastName\": \"Smithers\", \"etag\": {TestSetUp.ConcurrencyErrorETag} }}", id));
         }
@@ -400,9 +442,11 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void D130_Patch()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 4.ToGuid();
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
@@ -411,7 +455,7 @@ namespace My.Hr.Test
             v.EmergencyContacts = null;
 
             // Update the value.
-            v = AgentTester.Test<EmployeeAgent, Employee>()
+            v = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectChangeLogUpdated()
                 .ExpectETag(v.ETag)
@@ -421,7 +465,7 @@ namespace My.Hr.Test
                 .Run(a => a.PatchAsync(WebApiPatchOption.MergePatch, $"{{ \"lastName\": \"{v.LastName}\", \"emergencyContacts\": null }}", id, new WebApiRequestOptions { ETag = v.ETag })).Value;
 
             // Check the value was updated properly.
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(id));
@@ -434,6 +478,8 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void E110_Delete()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             var v = new Employee
             {
                 Email = "j.jones@org.com",
@@ -448,24 +494,24 @@ namespace My.Hr.Test
             };
 
             // Create an employee in the future.
-            v = AgentTester.Test<EmployeeAgent, Employee>()
+            v = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.Created)
                 .ExpectEvent("My.Hr.Employee", "Created")
                 .Run(a => a.CreateAsync(v)).Value;
 
             // Delete value.
-            AgentTester.Test<EmployeeAgent>()
+            agentTester.Test<EmployeeAgent>()
                 .ExpectStatusCode(HttpStatusCode.NoContent)
                 .ExpectEvent($"My.Hr.Employee", "Deleted")
                 .Run(a => a.DeleteAsync(v.Id));
 
             // Check value no longer exists.
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.GetAsync(v.Id));
 
             // Delete again (should still be successful as a Delete is idempotent); note there should be no corresponding event as nothing actually happened.
-            AgentTester.Test<EmployeeAgent>()
+            agentTester.Test<EmployeeAgent>()
                 .ExpectStatusCode(HttpStatusCode.NoContent)
                 .Run(a => a.DeleteAsync(v.Id));
         }
@@ -477,7 +523,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void F110_Terminate_NotFound()
         {
-            AgentTester.Test<EmployeeAgent, Employee>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.TerminateAsync(new TerminationDetail { Date = DateTime.Now, Reason = "RD" }, 404.ToGuid()));
         }
@@ -485,7 +533,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void F120_Terminate_MoreThanOnce()
         {
-            var r = AgentTester.Test<EmployeeAgent, Employee>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.BadRequest)
                 .ExpectErrorType(ErrorType.ValidationError, "An Employee can not be terminated more than once.")
                 .Run(a => a.TerminateAsync(new TerminationDetail { Date = DateTime.Now, Reason = "RD" }, 2.ToGuid()));
@@ -494,7 +544,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void F130_Terminate_BeforeStart()
         {
-            var r = AgentTester.Test<EmployeeAgent, Employee>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.BadRequest)
                 .ExpectErrorType(ErrorType.ValidationError, "An Employee can not be terminated prior to their start date.")
                 .Run(a => a.TerminateAsync(new TerminationDetail { Date = new DateTime(1999, 12, 31), Reason = "RD" }, 1.ToGuid()));
@@ -503,13 +555,15 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void F140_Terminate()
         {
-            var v = AgentTester.Test<EmployeeAgent, Employee?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(1.ToGuid())).Value!;
 
             v.Termination = new TerminationDetail { Date = Cleaner.Clean(DateTime.Now, DateTimeTransform.DateOnly), Reason = "RD" };
 
-            v = AgentTester.Test<EmployeeAgent, Employee>()
+            v = agentTester.Test<EmployeeAgent, Employee>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectChangeLogUpdated()
                 .ExpectETag(v.ETag)
@@ -518,7 +572,7 @@ namespace My.Hr.Test
                 .ExpectEvent($"My.Hr.Employee", "Terminated")
                 .Run(a => a.TerminateAsync(v.Termination, 1.ToGuid())).Value!;
 
-            AgentTester.Test<EmployeeAgent, Employee?>()
+            agentTester.Test<EmployeeAgent, Employee?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(1.ToGuid()));