+++ samples/My.Hr/My.Hr.Test/Apis/PerformanceReviewTest.cs
@@ -12,17 +12,19 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 
-namespace My.Hr.Test
+namespace My.Hr.Test.Apis
 {
     [TestFixture, NonParallelizable]
-    public class PerformanceReviewTest : UsingAgentTesterServer<Startup>
+    public class PerformanceReviewTest
     {
         #region Get
 
         [Test, TestSetUp]
         public void A110_Get_NotFound()
         {
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.GetAsync(404.ToGuid()));
         }
@@ -30,7 +32,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A110_Get()
         {
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .IgnoreChangeLog()
                 .IgnoreETag()
@@ -53,7 +57,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A210_GetByEmployeeId_NotFound()
         {
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByEmployeeIdAsync(4.ToGuid())).Value!;
 
@@ -65,7 +71,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A220_GetByEmployeeId()
         {
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByEmployeeIdAsync(2.ToGuid())).Value!;
 
@@ -78,7 +86,9 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void A220_GetByEmployeeId_Last()
         {
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReviewCollectionResult>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetByEmployeeIdAsync(2.ToGuid(), PagingArgs.CreateSkipAndTake(0, 1))).Value!;
 
@@ -95,6 +105,8 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void B110_Create()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             var v = new PerformanceReview
             {
                 Date = new DateTime(2020, 06, 15),
@@ -104,7 +116,7 @@ namespace My.Hr.Test
             };
 
             // Create value.
-            v = AgentTester.Test<PerformanceReviewAgent, PerformanceReview>()
+            v = agentTester.Test<PerformanceReviewAgent, PerformanceReview>()
                 .ExpectStatusCode(HttpStatusCode.Created)
                 .ExpectChangeLogCreated()
                 .ExpectETag()
@@ -116,7 +128,7 @@ namespace My.Hr.Test
             Assert.AreEqual(3.ToGuid(), v.EmployeeId);
 
             // Check the value was created properly.
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(v.Id));
@@ -129,13 +141,15 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C110_Update_NotFound()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(3.ToGuid())).Value!;
 
             // Try updating with an invalid identifier.
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.UpdateAsync(v, 404.ToGuid()));
         }
@@ -143,20 +157,22 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C120_Update_Concurrency()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 3.ToGuid();
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
             // Try updating the value with an invalid eTag (if-match).
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.UpdateAsync(v, id, new WebApiRequestOptions { ETag = TestSetUp.ConcurrencyErrorETag }));
 
             // Try updating the value with an invalid eTag.
             v.ETag = TestSetUp.ConcurrencyErrorETag;
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview>()
                 .ExpectStatusCode(HttpStatusCode.PreconditionFailed)
                 .Run(a => a.UpdateAsync(v, id));
         }
@@ -164,9 +180,11 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void C130_Update()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             // Get an existing value.
             var id = 3.ToGuid();
-            var v = AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            var v = agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id)).Value!;
 
@@ -174,7 +192,7 @@ namespace My.Hr.Test
             v.Notes += "X";
 
             // Update the value.
-            v = AgentTester.Test<PerformanceReviewAgent, PerformanceReview>()
+            v = agentTester.Test<PerformanceReviewAgent, PerformanceReview>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectChangeLogUpdated()
                 .ExpectETag(v.ETag)
@@ -184,7 +202,7 @@ namespace My.Hr.Test
                 .Run(a => a.UpdateAsync(v, id)).Value!;
 
             // Check the value was updated properly.
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue(_ => v)
                 .Run(a => a.GetAsync(id));
@@ -197,26 +215,28 @@ namespace My.Hr.Test
         [Test, TestSetUp]
         public void E110_Delete()
         {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
             var id = 3.ToGuid();
 
             // Get an existing value.
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .Run(a => a.GetAsync(id));
 
             // Delete value.
-            AgentTester.Test<PerformanceReviewAgent>()
+            agentTester.Test<PerformanceReviewAgent>()
                 .ExpectStatusCode(HttpStatusCode.NoContent)
                 .ExpectEvent($"My.Hr.PerformanceReview", "Deleted")
                 .Run(a => a.DeleteAsync(id));
 
             // Check value no longer exists.
-            AgentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
+            agentTester.Test<PerformanceReviewAgent, PerformanceReview?>()
                 .ExpectStatusCode(HttpStatusCode.NotFound)
                 .Run(a => a.GetAsync(id));
 
             // Delete again (should still be successful as a Delete is idempotent); note there should be no corresponding event as nothing actually happened.
-            AgentTester.Test<PerformanceReviewAgent>()
+            agentTester.Test<PerformanceReviewAgent>()
                 .ExpectStatusCode(HttpStatusCode.NoContent)
                 .Run(a => a.DeleteAsync(id));
         }