+++ src/Beef.Core/RefData/ReferenceDataBaseString.cs
@@ -7,25 +7,25 @@ using System.Diagnostics;
 namespace Beef.RefData
 {
     /// <summary>
-    /// Represents a <b>ReferenceData</b> base class with an <see cref="Id"/> <see cref="System.Type"/> of <see cref="System.Int32"/>.
+    /// Represents a <b>ReferenceData</b> base class with an <see cref="Id"/> <see cref="System.Type"/> of <see cref="System.String"/>.
     /// </summary>
     [DebuggerDisplay("Id = {Id}, Code = {Code}, Text = {Text}, IsActive={IsActive}, IsValid={IsValid}")]
     [JsonObject(MemberSerialization.OptIn)]
-    public abstract class ReferenceDataBaseInt : ReferenceDataBase, IIntIdentifier
+    public abstract class ReferenceDataBaseString : ReferenceDataBase, IStringIdentifier
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="ReferenceDataBaseInt"/> class.
+        /// Initializes a new instance of the <see cref="ReferenceDataBaseString"/> class.
         /// </summary>
-        public ReferenceDataBaseInt() : base(ReferenceDataIdTypeCode.Int32, 0) { }
+        public ReferenceDataBaseString() : base(ReferenceDataIdTypeCode.String, null!) { }
 
         /// <summary>
         /// Gets or sets the identifier.
         /// </summary>
         /// <remarks>Once set this value cannot be updated (it becomes immutable).</remarks>
         [JsonProperty("id", Order = 0)]
-        public new int Id
+        public new string? Id
         {
-            get { return base.Id == null || !(base.Id is int) ? 0 : (int)base.Id; }
+            get { return base.Id == null || !(base.Id is string @string) ? null : @string; }
             set { base.Id = value; }
         }
     }